requirements.txt
affine==2.4.0
attrs==25.3.0
certifi==2025.4.26
click==8.2.1
click-plugins==1.1.1
cligj==0.7.2
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.2
imagecodecs>=2023.1.23
imageio==2.37.0
kiwisolver==1.4.8
lazy_loader==0.4
matplotlib==3.10.3
networkx==3.5
numpy==2.3.0
opencv-python==4.11.0.86
packaging==25.0
pillow==11.2.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
rasterio==1.4.3
scikit-image==0.25.2
scipy==1.15.3
six==1.17.0
tifffile==2025.6.1
tqdm==4.67.1
